Modules are backed by SQLite.
One consequence of doing this is that `norm` becomes async. I think
that makes sense: computation is an effect, it may take a long time
and much effort, and so this leaks in to the type of `norm`.

The SQLite schema is degenerate right now but I plan to store more
data there, particularly for bots.